Create the SearchParameters class.
This class will be the interface for interacting with the search
parameters. It is supposed to have a coherent API to make it pleasant
for the user to add, remove, and modify the search parameters.

Sadly, I messed up some git stuff, which forced me to make a huge commit
instead of commitings changes one at a time.